APS :: Adicionado Sistema de Checkpoint

+++ Models/Local.cs
@@ -4,11 +4,11 @@ public class Local
 {
     public int Id { get; set; }
     public string Nome { get; set; }
-    public decimal LatitudeMin { get; set; }
-    public decimal LatitudeMax { get; set; }
-    public decimal LongitudeMin { get; set; }
-    public decimal LongitudeMax { get; set; }
-    public decimal LatitudeEntrada { get; set; }
-    public decimal LongitudeEntrada { get; set; }
+    public int LatitudeMin { get; set; }
+    public int LatitudeMax { get; set; }
+    public int LongitudeMin { get; set; }
+    public int LongitudeMax { get; set; }
+    public int LatitudeEntrada { get; set; }
+    public int LongitudeEntrada { get; set; }
     public bool LocalAcessivel { get; set; }
 }
